% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{D-class}
\alias{D-class}
\alias{D}
\title{\code{D} class}
\description{
A test class, subclass of test class \code{A} via \code{C}.
}
\details{
This class is part of a pair of classes for testing an issue with S4
method dispatch.
}
